feat: Added course image creating and updating,
added delete and update buttons to one course wrapper

+++ frontend/src/features/courses/components/CourseCard/CourseCard.tsx
@@ -50,12 +50,14 @@ const CourseCard: React.FC<Props> = ({ course }) => {
   };
 
   const openEditPage = () => {
-    void router.push(`/edit-course/${course._id}`);
+    void router.push(`/courses/edit/${course._id}`);
   };
 
   const handleDelete = async () => {
-    await dispatch(deleteCourse(course._id));
-    dispatch(fetchCourses());
+    if (window.confirm('Подтвердите удаление курса')) {
+      await dispatch(deleteCourse(course._id));
+      dispatch(fetchCourses());
+    }
   };
 
   return (
